+++ test/account/AccountERC7579.test.js
@@ -4,7 +4,7 @@ const { ERC4337Helper } = require('../helpers/erc4337');
 const { PackedUserOperation } = require('../helpers/eip712-types');
 
 const { shouldBehaveLikeAccountCore, shouldBehaveLikeAccountERC7579 } = require('./Account.behavior');
-const { shouldBehaveLikeERC7739Signer } = require('../utils/cryptography/ERC7739Signer.behavior');
+const { shouldBehaveLikeERC7739 } = require('../utils/cryptography/ERC7739.behavior');
 
 async function fixture() {
   // EOAs and environment
@@ -57,7 +57,7 @@ describe('AccountERC7579', function () {
   shouldBehaveLikeAccountCore();
   shouldBehaveLikeAccountERC7579();
 
-  describe('ERC7739Signer', function () {
+  describe('ERC7739', function () {
     beforeEach(async function () {
       this.mock = await this.mock.deploy();
       this.signTypedData ??= (async (...args) => {
@@ -67,6 +67,6 @@ describe('AccountERC7579', function () {
       }).bind(this.signer);
     });
 
-    shouldBehaveLikeERC7739Signer();
+    shouldBehaveLikeERC7739();
   });
 });